Add user creation on startup (#16)
* Add user creation on startup

* Map device ID to running process

* Add devices per user && refactor of the layer interfaces

* Make mapping bettween measurements and user

* Add device validation for a concrete user

* Adopt new cli API version

* Add all cli resources in an archive

* Add cli binaries to the generated .zip

* Unify the units in the sql script

* Make server config files generic so they can me mounted on deploy

* Adopt mailsender from the API

* Update binaries

* Remove direct depencies

* Add checksum generation

+++ pkg/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID        int64  `json:"id" mapstructure:"id"`
+	ID        int    `json:"id" mapstructure:"id"`
 	UserName  string `json:"user_name" mapstructure:"user_name"`
 	FirstName string `json:"first_name" mapstructure:"first_name"`
 	LastName  string `json:"last_name" mapstructure:"last_name"`
